requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
astor==0.8.1
attrs==23.2.0
bandit==1.7.9
certifi==2024.6.2
cffi==1.16.0
coverage==7.5.3
darglint==1.8.1
Django==5.0.6
django-minio-backend==3.6.0
djangorestframework==3.15.1
docutils==0.21.2
eradicate==2.3.0
flake8==7.1.0
flake8-bandit==4.1.1
flake8-broken-line==1.0.0
flake8-bugbear==24.4.26
flake8-commas==2.1.0
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-isort==6.1.1
flake8-quotes==3.4.0
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
minio==7.2.7
pbr==6.0.0
pep8-naming==0.13.3
pillow==10.3.0
psycopg2==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pycryptodome==3.20.0
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.1
restructuredtext-lint==1.4.0
rich==13.7.1
snowballstemmer==2.2.0
sqlparse==0.5.0
stevedore==5.2.0
typing_extensions==4.12.2
urllib3==2.2.1
wemake-python-styleguide==0.19.2
